Updates the tls certificate manager to no longer use config files, and simply use a postgresql database to store/retrieve certificates that are fully managed through the use of API requests, for which there is now an SDK in PHP.

+++ site/controllers/CertificatesController.php
@@ -17,25 +17,25 @@ class CertificatesController extends AbstractSlimController
         $app->get('/api/certs', function (Slim\Psr7\Request $request, Slim\Psr7\Response $response, $args) {
             $controller = new CertificatesController($request, $response, $args);
             return $controller->handleRequestToListCertificates();
-        })->addMiddleware(new MiddlewareAdminAuth());
+        })->addMiddleware(new MiddlewareTokenAuth());
 
         # create a new certificate bundle
         $app->post('/api/certs', function (Slim\Psr7\Request $request, Slim\Psr7\Response $response, $args) {
             $controller = new CertificatesController($request, $response, $args);
-            return $controller->hanldeRequestToCreateNewCertificateBundle();
-        })->addMiddleware(new MiddlewareAdminAuth());
+            return $controller->handleRequestToCreateNewCertificateBundle();
+        })->addMiddleware(new MiddlewareTokenAuth());
 
         # update certificate bundle
         $app->patch('/api/certs/{id:[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}}', function (Slim\Psr7\Request $request, Slim\Psr7\Response $response, $args) {
             $controller = new CertificatesController($request, $response, $args);
             return $controller->hanldeRequestToUpdateCertificateBundle($args['id']);
-        })->addMiddleware(new MiddlewareAdminAuth());
+        })->addMiddleware(new MiddlewareTokenAuth());
 
         # get a specific certificate bundle
         $app->get('/api/certs/{id:[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}}', function (Slim\Psr7\Request $request, Slim\Psr7\Response $response, $args) {
             $controller = new CertificatesController($request, $response, $args);
             return $controller->handleRequestForCertificateBundle($args['id']);
-        });
+        })->addMiddleware(new MiddlewareTokenAuth());
 
         # delete a certificate bundle
         $app->delete('/api/certs/{id:[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}}', function (Slim\Psr7\Request $request, Slim\Psr7\Response $response, $args) {
@@ -47,7 +47,8 @@ class CertificatesController extends AbstractSlimController
 
     private function handleRequestToListCertificates()
     {
-        $certificates = CertificateBundleTable::getInstance()->loadAll();
+        $authToken = Auth::getAuthToken($this->m_request);
+        $certificates = CertificateBundleTable::getInstance()->fetchForAuthTokenReadAccess($authToken);
         $responseData = [];
 
         foreach ($certificates as $certificate)
@@ -80,13 +81,18 @@ class CertificatesController extends AbstractSlimController
         {
             /* @var $certificateBundle CertificateBundleRecord */
             $certificateBundle = CertificateBundleTable::getInstance()->load($certificateBundleId);
+            $authToken = Auth::getAuthToken($this->m_request);
 
-            /* @TODO - authenticate auth token against certificate */
-            $this->authenticateCertificateRequest($certificateBundle);
+            if (!$authToken->hasReadCertificatePermission($certificateBundle))
+            {
+                throw new ExceptionPermissionDenied("You do not have permission to read this certificate.");
+            }
 
             $responseData = [
                 'id' => $certificateBundle->getId(),
                 'name' => $certificateBundle->getName(),
+                'cert' => $certificateBundle->getCert(),
+                'chain' => $certificateBundle->getChain(),
                 'fullchain' => $certificateBundle->getFullchain(),
                 'private_key' => $certificateBundle->getPrivateKey(),
             ];
@@ -103,91 +109,8 @@ class CertificatesController extends AbstractSlimController
 
             $response = SlimLib::createJsonResponse($responseData, HttpCode::INTERNAL_SERVER_ERROR);
         }
-        catch (ExceptionBadRequest $badRequest)
-        {
-            $response = SlimLib::createJsonResponse(
-                ["error" => ["message" => $badRequest->getMessage()]],
-                HttpCode::BAD_REQUEST
-            );
-        }
-        catch (ExceptionUnauthorized)
-        {
-            $response = SlimLib::createJsonResponse(
-                ["error" => ["message" => "Authentication failed"]],
-                HttpCode::UNAUTHORIZED
-            );
-        }
-        catch (Exception)
-        {
-            $responseData = [
-                "error" => [
-                    "message" => "Whoops! Something went wrong. Please try again or contact support.",
-                ]
-            ];
-
-            $response = SlimLib::createJsonResponse($responseData, HttpCode::INTERNAL_SERVER_ERROR);
-        }
 
         return $response;
-    }
-
-
-    /**
-     * @param CertificateBundleRecord $certificate
-     * @return void
-     * @throws ExceptionBadRequest
-     * @throws ExceptionUnauthorized - if the user is is not authorized to fetch the specified certificate.
-     * @throws \Programster\PgsqlLib\Exceptions\ExceptionQueryError
-     * @throws \Programster\PgsqlLib\Exceptions\ExceptionUnexpectedValueType
-     */
-    private function authenticateCertificateRequest(CertificateBundleRecord $certificate)
-    {
-        $authHeaders = $this->m_request->getHeader('Authorization');
-
-        if (count($authHeaders) === 0)
-        {
-            throw new ExceptionBadRequest("Missing required authorization header.");
-        }
-
-        $authHeader = $authHeaders[0];
-
-        if (str_contains($authHeader, "Bearer") === false)
-        {
-            throw new ExceptionBadRequest("Authorization header needs to be a bearer token.");
-        }
-
-        $encodedBearerToken = trim(str_replace('Bearer', '', $authHeader));
-        $bearerToken = base64_decode($encodedBearerToken);
-
-
-        if (password_verify($bearerToken, ADMIN_AUTH_TOKEN_HASH))
-        {
-            $authorized = true;
-        }
-        else
-        {
-            // for this specific route, one can use an assigned auth token, check against these.
-            $authTokens = $certificate->fetchAuthTokens();
-            $authorized = false;
-
-            if (count($authTokens) > 0)
-            {
-                foreach ($authTokens as $authToken)
-                {
-                    /* @var $authToken AuthTokenRecord */
-                    if (password_verify($bearerToken, $authToken->getTokenHash()))
-                    {
-                        $authorized = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if (!$authorized)
-        {
-            throw new ExceptionUnauthorized();
-        }
     }
 
 
@@ -195,14 +118,30 @@ class CertificatesController extends AbstractSlimController
     {
         try
         {
-            $allPostFields = $this->m_request->getParsedBody();
-
-            /* @var $certificateBundleRecord CertificateBundleRecord */
+            /* @var $certificateBundle CertificateBundleRecord */
             $certificateBundle = CertificateBundleTable::getInstance()->load($certificateBundleId);
+            $authToken = Auth::getAuthToken($this->m_request);
+
+            if (!$authToken->hasUpdateCertificatePermission($certificateBundle))
+            {
+                throw new ExceptionPermissionDenied("You do not have permission to update this certificate.");
+            }
+
+            $allPostFields = $this->m_request->getParsedBody();
 
             if (array_key_exists('name', $allPostFields))
             {
                 $certificateBundle->setName($allPostFields['name']);
+            }
+
+            if (array_key_exists('cert', $allPostFields))
+            {
+                $certificateBundle->setCert($allPostFields['cert']);
+            }
+
+            if (array_key_exists('chain', $allPostFields))
+            {
+                $certificateBundle->setChain($allPostFields['chain']);
             }
 
             if (array_key_exists('fullchain', $allPostFields))
@@ -243,13 +182,22 @@ class CertificatesController extends AbstractSlimController
     }
 
 
-    private function hanldeRequestToCreateNewCertificateBundle()
+    private function handleRequestToCreateNewCertificateBundle()
     {
         try
         {
+            $authToken = Auth::getAuthToken($this->m_request);
+
+            if (in_array($authToken->getAccessLevel(), [AuthTokenLevel::ADMIN, AuthTokenLevel::CERTIFICATE_CREATOR]) === false)
+            {
+                throw new ExceptionPermissionDenied("You need to have an admin or creator token to create certificates.");
+            }
+
             $requiredPostFields = [
                 'id',
                 'name',
+                'cert',
+                'chain',
                 'fullchain',
                 'private_key',
             ];
@@ -270,19 +218,33 @@ class CertificatesController extends AbstractSlimController
                 throw new ExceptionModelAlreadyExists("A certificate bundle with that ID already exists.");
             }
 
-            if (count(CertificateBundleTable::getInstance()->loadWhereAnd(['name' => $allPostFields['name']])) > 0)
+            // validate ID is a UUID
+            $id = $allPostFields['id'];
+            $uuidPattern = '/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i';
+
+            if (preg_match($uuidPattern, $id) === false)
             {
-                throw new ExceptionModelAlreadyExists("A certificate bundle with that name already exists.");
+                throw new ExceptionBadRequest("The ID needs to be a UUIDv4.");
             }
 
             $certificateBundleRecord = CertificateBundleRecord::createNew(
                 $allPostFields['id'],
                 $allPostFields['name'],
+                $allPostFields['cert'],
+                $allPostFields['chain'],
                 $allPostFields['fullchain'],
                 $allPostFields['private_key'],
             );
 
             $certificateBundleRecord->save();
+
+            // only create assignment records if the token is NOT an admin token or full reader because
+            // a) this has no effect on those tokens abilities.
+            // b) the admin token from env file wont exist in the database, so FKs will fail for insertion.
+            if ($authToken->isAdmin() === false && $authToken->isFullReader() === false)
+            {
+                AuthTokenAssignmentRecord::createNew($authToken, $certificateBundleRecord)->save();
+            }
 
             $response = SlimLib::createJsonResponse(['message' => "Certificate bundle created."], HttpCode::CREATED);
         }
@@ -291,16 +253,6 @@ class CertificatesController extends AbstractSlimController
             $responseData = [
                 "error" => [
                     "message" => $passthruException->getMessage(),
-                ]
-            ];
-
-            $response = SlimLib::createJsonResponse($responseData, HttpCode::INTERNAL_SERVER_ERROR);
-        }
-        catch (Exception $e)
-        {
-            $responseData = [
-                "error" => [
-                    "message" => "Whoops! Something went wrong. Please try again or contact support.",
                 ]
             ];
 
@@ -315,7 +267,15 @@ class CertificatesController extends AbstractSlimController
     {
         try
         {
+            $authToken = Auth::getAuthToken();
+            /* @var $certificateBundle CertificateBundleRecord */
             $certificateBundle = CertificateBundleTable::getInstance()->load($certificatesBundleId);
+
+            if ($authToken->hasDeleteCertificatePermission($certificateBundle) === false)
+            {
+                throw new ExceptionPermissionDenied("You do not have permission to delete that certificate.");
+            }
+
             $certificateBundle->delete();
 
             $newResponse = SlimLib::createJsonResponse(
